html/requirements.txt
Django
PIL
Pillow
argparse
django-oembed
django-taggit
easy-thumbnails
six
wsgiref
yolk
